moved counter to a separate page
